Fixed #137 remove DefineAllMeasures menu from measure
And moved metadata methods out of ToolPaneBaseViewModel into MetadataPaneViewModel
