+++ installer/DaxStudio.iss
@@ -27,7 +27,11 @@
 ; (To generate a new GUID, click Tools | Generate GUID inside the IDE.)
 ;AppId={{2DDF2B91-978D-4B47-AF1C-15E6C07ADEAD}
 AppId={{CE2CEA93-9DD3-4724-8FE3-FCBF0A0915C1}
-AppName={#MyAppName}
+#ifdef Preview
+AppName={#MyAppName} {#myAppMajor}.{#myAppMinor}.{#myAppRevision} ({#Preview})
+#else
+AppName={#MyAppName} {#myAppMajor}.{#myAppMinor}.{#myAppRevision}
+#endif
 AppVersion={#MyAppVersion}
 AppVerName={#MyAppName} {#MyAppVersion}
 AppPublisher={#MyAppPublisher}
